move token source provider logic to admin to avoid circular deps
