+++ clients/go/admin/client_test.go
@@ -181,9 +181,11 @@ func TestGetAuthenticationDialOptionPkce(t *testing.T) {
 		mockAuthClient := new(mocks.AuthMetadataServiceClient)
 		mockTokenCache.OnGetTokenMatch().Return(&tokenData, nil)
 		mockTokenCache.OnSaveTokenMatch(mock.Anything).Return(nil)
+		tokenSourceProvider, err := NewTokenSourceProvider(ctx, adminServiceConfig, mockTokenCache, mockAuthClient)
+		assert.NotNil(t, err)
 		mockAuthClient.OnGetOAuth2MetadataMatch(mock.Anything, mock.Anything).Return(metatdata, nil)
 		mockAuthClient.OnGetPublicClientConfigMatch(mock.Anything, mock.Anything).Return(clientMetatadata, nil)
-		dialOption, err := getAuthenticationDialOption(ctx, adminServiceConfig, mockTokenCache, mockAuthClient)
+		dialOption, err := getAuthenticationDialOption(ctx, adminServiceConfig, tokenSourceProvider, mockAuthClient)
 		assert.NotNil(t, dialOption)
 		assert.Nil(t, err)
 	})
@@ -193,9 +195,11 @@ func TestGetAuthenticationDialOptionPkce(t *testing.T) {
 		mockAuthClient := new(mocks.AuthMetadataServiceClient)
 		mockTokenCache.OnGetTokenMatch().Return(&tokenData, nil)
 		mockTokenCache.OnSaveTokenMatch(mock.Anything).Return(nil)
+		tokenSourceProvider, err := NewTokenSourceProvider(ctx, adminServiceConfig, mockTokenCache, mockAuthClient)
+		assert.NotNil(t, err)
 		mockAuthClient.OnGetOAuth2MetadataMatch(mock.Anything, mock.Anything).Return(metatdata, nil)
 		mockAuthClient.OnGetPublicClientConfigMatch(mock.Anything, mock.Anything).Return(clientMetatadata, nil)
-		dialOption, err := getAuthenticationDialOption(ctx, adminServiceConfig, mockTokenCache, mockAuthClient)
+		dialOption, err := getAuthenticationDialOption(ctx, adminServiceConfig, tokenSourceProvider, mockAuthClient)
 		assert.Nil(t, dialOption)
 		assert.NotNil(t, err)
 	})
@@ -231,7 +235,7 @@ func Test_getPkceAuthTokenSource(t *testing.T) {
 		assert.NoError(t, err)
 
 		http.DefaultServeMux = http.NewServeMux()
-		dialOption, err := getPkceAuthTokenSource(ctx, orchestrator)
+		dialOption, err := GetPKCEAuthTokenSource(ctx, orchestrator)
 		assert.Nil(t, dialOption)
 		assert.Error(t, err)
 	})
